+++ eduflex-manage/src/main/java/com/eduflex/manage/goal/domain/dto/GoalStudentDto.java
@@ -3,12 +3,11 @@ package com.eduflex.manage.goal.domain.dto;
 import com.eduflex.common.core.domain.BaseEntity;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class GoalStudentDto extends BaseEntity {
     private Long userId;
     private Long status;
-    private List<BigDecimal> progress;
+    private List<Integer> progress;
 }
