+++ eduflex-manage/src/main/java/com/eduflex/manage/student/domain/StudentCourse.java
@@ -1,18 +1,17 @@
 package com.eduflex.manage.student.domain;
 
-import java.math.BigDecimal;
-import java.util.Date;
-
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.eduflex.common.annotation.Excel;
+import com.eduflex.common.core.domain.BaseEntity;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import com.eduflex.common.annotation.Excel;
-import com.eduflex.common.core.domain.BaseEntity;
+
+import java.util.Date;
 
 /**
  * 学生选课对象 tb_student_course
@@ -51,7 +50,7 @@ public class StudentCourse extends BaseEntity {
      * 学习进度(%)
      */
     @Excel(name = "学习进度(%)")
-    private BigDecimal progress;
+    private Integer progress;
 
     /**
      * 状态（0-未开始 1-进行中 2-已完成）
